Update architecture diagram files
Remove the PlantUML C2 diagram and replace it with the Draw.io version.

Remove the /out folder and place the C1 and C2 source files into a new
folder called /src.

Create separate diagrams for the total architecture (all services blue)
and the diagram indicating the group assignment (services colored based
on the group they were assigned to). The former diagram is more useful
for the next team working on the project.

Export C2 diagram (general and group assignment) in svg and pdf format.
The former is great for display in MarkDown files later, and the latter
is a good format to share or print the diagrams.